% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borr.R
\docType{data}
\name{borr}
\alias{borr}
\title{Borrelioositartuntojen jakautuminen Suomessa 2017}
\format{
A data frame with 6 rows and 2 variables:
\describe{
\item{alue}{Region, sairaanhoitopiiri}
\item{pros}{Percent}
...
}
}
\source{
\url{https://bit.ly/THL-esimerkkiaineistot}
}
\usage{
borr
}
\description{
Borrelioositartuntojen jakautuminen Suomessa 2017
}
\keyword{datasets}
