% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thlAnnotate.R
\name{thlAnnotate}
\alias{thlAnnotate}
\title{Add annotations to the plot}
\usage{
thlAnnotate(
  type = c("box", "label"),
  x,
  y,
  xend = x,
  yend = y,
  label,
  arrow.size = 10,
  text.size = 16,
  hjust = 0.5,
  style = c("white", "dark"),
  fill = TRUE,
  boldFont = TRUE,
  capitalized = FALSE,
  box.size = 0.4
)
}
\arguments{
\item{type}{Type of annotation wanted: "box" for a text box or "label" for text}

\item{x}{x-axis location for annotation}

\item{y}{y-axis location for annotation}

\item{xend}{location on the x-axis for the tip of the arrow, defaults to x}

\item{yend}{location on the x-axis for the tip of the arrow, defaults to y}

\item{label}{text to be annotated}

\item{arrow.size}{size of the arrow head in PostScript points}

\item{text.size}{size of text head in PostScript points}

\item{hjust}{horizontal justification of the text, 0 = left justified, 0 = right justified. Defaults to 0.5.}

\item{style}{color style of the annotation ("white" or "dark")}

\item{fill}{Fill the background of the annotation with color? Defaults to TRUE}

\item{boldFont}{set TRUE (default) for bolded font}

\item{capitalized}{set TRUE for upper case letters, default is FALSE}

\item{box.size}{unused}
}
\description{
Add annotations to the plot
}
\author{
Salla Toikkanen, Tarja Palosaari, Petteri Mäntymaa
}
