% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thlNaLines.R
\name{thlNaLines}
\alias{thlNaLines}
\title{Linear interpolation of missing values with dotted line in line plots}
\usage{
thlNaLines(data, xvar, yvar, groupvar)
}
\arguments{
\item{data}{dataset}

\item{xvar}{x axis variable}

\item{yvar}{y thlNaLines}

\item{groupvar}{omit setting NULL}
}
\description{
Linear interpolation of missing values with dotted line in line plots
}
\author{
Salla Toikkanen,  Tarja Palosaari, Petteri Mäntymaa
}
