% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thlPtsConvert.R
\name{thlPtsConvert}
\alias{thlPtsConvert}
\title{Conversion from PostScript points to millimeters}
\usage{
thlPtsConvert(pts)
}
\arguments{
\item{pts}{number of PostScript points}
}
\value{
Numeric value
}
\description{
Conversion from PostScript points to millimeters
}
\author{
Salla Toikkanen, Tarja Palosaari, Petteri Mäntymaa
}
