% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thlShade.R
\name{thlShade}
\alias{thlShade}
\title{Transparent or lighter versions of colors}
\usage{
thlShade(col, alpha = 0.25, thin = 1)
}
\arguments{
\item{col}{A color. Eighter a name such as "red" or preferrably a hex string
such as "#FF0000"}

\item{alpha}{Amount of transparency, 1=not transparent, 0=fully transparent.
Recycled.}

\item{thin}{Amount of lightening, 1=none, 0=white.}
}
\value{
A vector of rgba values suitable for using in graphics.
}
\description{
For each color, create a lighter version either by introducing transparency
or by just dimming down.
}
\details{
Same thin is used for all components of RGB. Thus \code{thin=0} always
results in white. No value checking is done, so either \code{thin} or
\code{alpha} outside 0,1 most likely results in error.
}
\note{
Not all graphics devices support transparency. Parameter \code{thin}
may be helpful in these situations.
}
\examples{

thlShade("red", alpha = .40)
thlShade(thlColors(type = "main"), alpha = .8, thin = .5)


}
\seealso{
\code{\link{col2rgb}}, \code{\link{rgb}}
}
\author{
Mikko J Virtanen, Petteri Mäntymaa, Tarja Palosaari, Salla Toikkanen
}
\keyword{Graphics}
