% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thlTheme.R
\name{thlTheme}
\alias{thlTheme}
\title{THL visual theme using ggplot2}
\usage{
thlTheme(
  show.grid.y = TRUE,
  show.grid.x = FALSE,
  base.size = 16,
  basic.lwd = 3,
  legend.position = "none",
  horizontal = FALSE,
  x.axis.title = FALSE
)
}
\arguments{
\item{show.grid.y}{Show background horizontal grid TRUE/FALSE?}

\item{show.grid.x}{Show background vertical grid TRUE/FALSE?}

\item{base.size}{Base size for elements in PostScript points}

\item{basic.lwd}{Basic linewidth in PostScript points}

\item{legend.position}{Position of the legend text}

\item{horizontal}{Set as TRUE if plotting horizontal bars}

\item{x.axis.title}{Set as TRUE if x-axis label is desired}
}
\value{
theme() object, to be combined with ggplot object with "+"
}
\description{
THL visual theme using ggplot2
}
\author{
Tarja Palosaari,  Salla Toikkanen, Petteri Mäntymaa
}
