% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thlYaxisControl.R
\name{thlYaxisControl}
\alias{thlYaxisControl}
\title{Control y-axis limits, big marks and decimal marks}
\usage{
thlYaxisControl(
  lang = "fi",
  limits = NULL,
  breaks = waiver(),
  sec.axis = FALSE,
  sec.axis.breaks = waiver(),
  sec.axis.labels = waiver()
)
}
\arguments{
\item{lang}{Language: "fi" (default) for Finnish or "en" for English}

\item{limits}{Specify y-axis limits}

\item{breaks}{waiver() is the default breaks, see scale_continuous()}

\item{sec.axis}{specify a secondary axis, see scale_continuous()}

\item{sec.axis.breaks}{waiver() is the default breaks for a secondary axis}

\item{sec.axis.labels}{the labels to be plotted on secondary axis,
waiver() is the default, see sec_axis()}
}
\value{
ggplot options, to be combined with ggplot object with "+"
}
\description{
Control y-axis limits, big marks and decimal marks
}
\seealso{
\code{\link{scale_x_continuous}} \code{\link{sec_axis}}
}
\author{
Salla Toikkanen, Tarja Palosaari, Petteri Mäntymaa
}
