% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yli180.R
\docType{data}
\name{yli180}
\alias{yli180}
\title{Erikoissairaanhoidosssa yli 180 vrk odottaneet}
\format{A data frame with 10 rows and 2 variables:
\describe{
\item{kk}{month/year}
\item{pros}{percent}
...
}}
\source{
\url{https://bit.ly/THL-esimerkkiaineistot}
}
\usage{
yli180
}
\description{
Käytössä olevat rajoitukset
Palveluntuottaja	SAIRAANHOITOPIIRIT
Erikoisala	KIRURGIA
Mittari	Yli 180 vrk, \%, hoitoa odottaneet
Kuutio päivitetty 27.05.2018
(c) Terveyden ja hyvinvoinnin laitos 2018, CC BY 4.0
}
\keyword{datasets}
