% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yli90.R
\docType{data}
\name{yli90}
\alias{yli90}
\title{Perusterveydenhuollossa hoitoonpääsyä yli 90 vrk odottaneiden määrä}
\format{A data frame with 6 rows and 7 variables:
\describe{
\item{AVI}{Region, AVI}
\item{1/2018}{First time point}
\item{2/2018}{Second time point}
\item{3/2018}{Third time point}
\item{4/2018}{Fourth time point}
\item{5/2018}{Fifth time point}
\item{6/2018}{Sixth time point}
...
}}
\source{
\url{https://bit.ly/THL-esimerkkiaineistot}
}
\usage{
yli90
}
\description{
Perusterveydenhuollossa hoitoonpääsyä yli 90 vrk odottaneiden määrä
}
\keyword{datasets}
